Add use-application-dns.net
The DNS response for this domain is used to test if DoH should be
enabled or not. We're setting up a redirect in case someone is curious
and tries to access it in their browser. For more details, see
https://jira.mozilla.com/browse/SE-1564

This is being migrated from the datacenter redirects cluster to
refractr.