feature(add-headers): Add additional headers for particular path
Bugzill https://bugzilla.mozilla.org/show_bug.cgi?id=1747089 is
requesting that we provided additional headers for a particular path
so that the Matrix client can correctly grab the client configuration.

This PR adds that path with the behavior requested and adds a new
refractr "complex type" which allows adding additional headers.

I had to select a different name for the key `endpoint` instead of
`redirect` as the other schema can match on as little as 1 matching
value and we would then pick the incorrect schema during compilation.

The additional headers from the main block have to be repeated,
otherwise nginx will assume we are overriding the headers and not
include them. This is expected behavior for nginx.

Lastly, the fuzzy match is required otherwise refractr renders a `/`
as a `Location = /` which nginx interprets as matching exactly `/`
where as a `Location /` which match the fall through. This is probably
introducing unexpected and undesired behavior in other places.